Split into new trait HasFormActions
